fixed average pressure bug

+++ code/Task4/alpotential.h
@@ -14,6 +14,6 @@ extern double get_virial_AL(double[][3], double, int);
 extern double instantaneous_temperature(double,int);
 extern double averaged_temperature(double*,int,double,int);
 extern double instantaneous_pressure(double,double,int,double);
-extern double averaged_pressure(double*,double*,double,int,int);
+extern double averaged_pressure(double*,double*,double,double,int);
 
 #endif
